+++ examples/discord_cog.py
@@ -16,12 +16,10 @@ class BrawlStars(commands.Cog, name='Brawl Stars'):
         try:
             player = await self.client.get_profile(tag)
         except brawlstats.RequestError as e:  # catches all exceptions
-            # sends code and error message
-            return await ctx.send('```\n{}: {}\n```'.format(e.code, e.message))
+            return await ctx.send('```\n{}: {}\n```'.format(e.code, e.message))  # sends code and error message
         em = discord.Embed(title='{0.name} ({0.tag})'.format(player))
 
-        # you could make this better by using embed fields
-        em.description = 'Trophies: {}'.format(player.trophies)
+        em.description = 'Trophies: {}'.format(player.trophies)  # you could make this better by using embed fields
         await ctx.send(embed=em)