+++ tests/test_blocking.py
@@ -19,9 +19,6 @@ class TestBlockingClient(unittest.TestCase):
             base_url=os.getenv('base_url')
         )
 
-    def tearDown(self):
-        self.client.close()
-
     def test_get_player(self):
         player = self.client.get_player(self.PLAYER_TAG)
         self.assertIsInstance(player, brawlstats.Player)
@@ -31,12 +28,9 @@ class TestBlockingClient(unittest.TestCase):
         self.assertIsInstance(club, brawlstats.Club)
         self.assertEqual(club.tag, self.CLUB_TAG)
 
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_player, '2PPPPPPP')
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_player, 'P')
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_player, 'AAA')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_player, '2PPPPPPP')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_player, 'P')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_player, 'AAA')
 
     def test_get_battle_logs(self):
         battle_logs = self.client.get_battle_logs(self.PLAYER_TAG)
@@ -51,52 +45,40 @@ class TestBlockingClient(unittest.TestCase):
         self.assertIsInstance(club_members, brawlstats.Members)
         self.assertIn(self.PLAYER_TAG, [x.tag for x in club_members])
 
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_club, '8GGGGGGG')
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_club, 'P')
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_club, 'AAA')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_club, '8GGGGGGG')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_club, 'P')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_club, 'AAA')
 
     def test_get_club_members(self):
         club_members = self.client.get_club_members(self.CLUB_TAG)
         self.assertIsInstance(club_members, brawlstats.Members)
         self.assertIn(self.PLAYER_TAG, [x.tag for x in club_members])
 
-        self.assertRaises(
-            brawlstats.NotFoundError, self.client.get_club_members, '8GGGGGGG')
+        self.assertRaises(brawlstats.NotFoundError, self.client.get_club_members, '8GGGGGGG')
 
     def test_get_rankings(self):
         player_ranking = self.client.get_rankings(ranking='players')
         self.assertIsInstance(player_ranking, brawlstats.Ranking)
 
-        us_player_ranking = self.client.get_rankings(
-            ranking='players', region='US', limit=1)
+        us_player_ranking = self.client.get_rankings(ranking='players', region='US', limit=1)
         self.assertIsInstance(us_player_ranking, brawlstats.Ranking)
         self.assertTrue(len(us_player_ranking) == 1)
 
-        self.assertRaises(
-            ValueError, self.client.get_rankings, ranking='people')
-        self.assertRaises(
-            ValueError, self.client.get_rankings, ranking='people', limit=0)
-        self.assertRaises(
-            ValueError, self.client.get_rankings,
-            ranking='brawlers', brawler='SharpBit')
+        self.assertRaises(ValueError, self.client.get_rankings, ranking='people')
+        self.assertRaises(ValueError, self.client.get_rankings, ranking='people', limit=0)
+        self.assertRaises(ValueError, self.client.get_rankings, ranking='brawlers', brawler='SharpBit')
 
         club_ranking = self.client.get_rankings(ranking='clubs')
         self.assertIsInstance(club_ranking, brawlstats.Ranking)
 
-        us_club_ranking = self.client.get_rankings(
-            ranking='clubs', region='US', limit=1)
+        us_club_ranking = self.client.get_rankings(ranking='clubs', region='US', limit=1)
         self.assertIsInstance(us_club_ranking, brawlstats.Ranking)
         self.assertTrue(len(us_club_ranking) == 1)
 
-        brawler_ranking = self.client.get_rankings(
-            ranking='brawlers', brawler='Shelly')
+        brawler_ranking = self.client.get_rankings(ranking='brawlers', brawler='Shelly')
         self.assertIsInstance(brawler_ranking, brawlstats.Ranking)
 
-        us_brawler_ranking = self.client.get_rankings(
-            ranking='brawlers', brawler=16000000, region='US', limit=1)
+        us_brawler_ranking = self.client.get_rankings(ranking='brawlers', brawler=16000000, region='US', limit=1)
         self.assertIsInstance(us_brawler_ranking, brawlstats.Ranking)
         self.assertTrue(len(us_brawler_ranking) == 1)
 
@@ -112,6 +94,9 @@ class TestBlockingClient(unittest.TestCase):
     def test_get_brawlers(self):
         brawlers = self.client.get_brawlers()
         self.assertIsInstance(brawlers, brawlstats.Brawlers)
+
+    def tearDown(self):
+        self.client.close()
 
 
 if __name__ == '__main__':