+++ brawlstats/models.py
@@ -5,6 +5,24 @@ from .utils import bstag, find_brawler
 __all__ = ['Player', 'Club', 'Members', 'Ranking', 'BattleLog', 'Brawlers']
 
 
+# for child classes
+class MyBox(Box):
+    def __getattr__(self, attr):
+        try:
+            return super().__getattr__(attr)
+        except AttributeError:
+            return None  # users can use an if statement rather than try/except to find a missing attribute
+
+
+# for child classes
+class MyBoxList(BoxList):
+    def __getattr__(self, attr):
+        try:
+            return super().__getattr__(attr)
+        except AttributeError:
+            return None  # users can use an if statement rather than try/except to find a missing attribute
+
+
 class BaseBox:
     def __init__(self, client, data):
         self.client = client
@@ -12,7 +30,7 @@ class BaseBox:
 
     def from_data(self, data):
         self.raw_data = data
-        self._boxed_data = Box(data, camel_killer_box=True)
+        self._boxed_data = MyBox(data, camel_killer_box=True)
         return self
 
     def __getattr__(self, attr):
@@ -35,7 +53,7 @@ class BaseBoxList(BaseBox):
     def from_data(self, data):
         data = data['items']
         self.raw_data = data
-        self._boxed_data = BoxList(data, camel_killer_box=True)
+        self._boxed_data = MyBoxList(data, box_class=MyBox, camel_killer_box=True)
         return self
 
     def __len__(self):
@@ -128,16 +146,18 @@ class Brawlers(BaseBoxList):
 
     def find(self, pattern, match):
         """
-        Find brawler containing template
+        Find first match brawler containing template
 
         Parameters
         ----------
-        pattern: any python instance, usually str or int
+        pattern: Any, usually str or int
             `match` value to find in brawlers
 
-        match: any python instance, usually str or int
+        match: Any, usually str or int
             key by which the search will be performed
 
         Returns brawler object
         """
+        if pattern == "name" and isinstance(match, str):
+            match = match.upper()
         return find_brawler(self, pattern, match)