+++ brawlstats/__init__.py
@@ -2,13 +2,12 @@ from .core import Client
 from .models import *
 from .errors import *
 
-
 ############
 # METADATA #
 ############
 
 
-__version__ = 'v4.0.3'
+__version__ = 'v4.0.4'
 __title__ = 'brawlstats'
 __license__ = 'MIT'
 __author__ = 'SharpBit'